+++ pkg/core/core.go
@@ -18,6 +18,11 @@ import (
 func PreRun() {
 	os.RemoteNode = &conn.Node{IP: os.RemoteNodeIP}
 	if IsoPath != "" {
+		err := conn.CheckSSHConnection(&conn.Node{IP: os.RemoteNode.IP})
+		if err != nil {
+			os.Exit(err.Error(), 1)
+		}
+		os.AddToSudoers(os.RemoteNode.IP)
 		util.StartSpinner(fmt.Sprintf("Umounting previous iso dir \"%s\" if exists on \"%s\"", constant.IsoMountDir, os.RemoteNode.IP))
 		os.UmountISO(constant.IsoMountDir, os.RemoteNode.IP)
 		util.StopSpinner("", logsymbols.Success)
@@ -28,7 +33,7 @@ func PreRun() {
 		fileStr := os.RunCommand(fmt.Sprintf("sudo cat %s/versions", constant.IsoMountDir), true)
 		println(fileStr)
 		var isoVersions model.IsoVersions
-		err := yaml.Unmarshal([]byte(fileStr), &isoVersions)
+		err = yaml.Unmarshal([]byte(fileStr), &isoVersions)
 		if err != nil {
 			os.Exit(err.Error(), 1)
 		}
@@ -37,6 +42,7 @@ func PreRun() {
 		CalicoVersion = isoVersions.Calico
 		EtcdVersion = isoVersions.Etcd
 		HelmVersion = isoVersions.Helm
+		HelmfileVersion = isoVersions.Helmfile
 	}
 	helmSemVer, _ := version.NewVersion(HelmVersion)
 	minHelmVer, _ := version.NewVersion("3.0.0")